+++ changelog.md
@@ -4,3 +4,4 @@
 - Chests now extend ChestBlocks instead of AbstractChestBlocks. This should improve compatibility and reduce the amount of crashes.
 - Chests screens no longer force scrolling when leaking by an extra row (thanks glisco!)
 - Chest recipes now use the `c:wooden_chests` tag for crafting recipes
+- Added a block set for Hydrargym (Storage Block, Anvil, nuggets)